ci: tighten phpstan defaults + wire pre-commit + commit-msg hooks
Three hardening steps that normalize the app's CI gate against the
conventions already used by pyrobase.

## phpstan.neon.dist: removed two lenient overrides

The agent that landed PHPStan initially set:

    treatPhpDocTypesAsCertain: false
    reportUnmatchedIgnoredErrors: false

Both are now removed so PHPStan uses its defaults (both `true`):

- `treatPhpDocTypesAsCertain: true` — PHPDoc-declared types behave like
  real types for always-true/always-false detection. The `false`
  override was masking real dead checks.
- `reportUnmatchedIgnoredErrors: true` — stale `@phpstan-ignore`
  comments fail the build. We don't use any, but enforcing the
  absence keeps it that way.

## 19 dead checks surfaced + fixed

The tightening surfaced 19 dead checks the loose flags had been
hiding. Mostly `nullCoalesce.offset` (?? defaults on guaranteed-
present array keys) and `function.alreadyNarrowedType`. Two notable
judgement calls:

- `class-remote-manager.php:73` + `class-request-builder.php:544` —
  the docblocks claimed tighter types than the runtime actually
  guarantees (WP filter callback / LruCache callback boundaries).
  Loosened to `@param mixed` since the defensive `is_array` /
  `instanceof \stdClass` checks in the bodies ARE the real boundary
  validation, not dead code.
- `class-partition-reader.php:65-66` — the docblock said
  `array{segment_id:int, offset:int}` but a unit test deliberately
  passes `[]` to verify the defensive defaults. Relaxed to
  `array{segment_id?:int, offset?:int}` so the `?? 0` fallbacks are
  type-honest.

Full fix list:

1. class-admin.php:152 — booleanAnd.leftAlwaysTrue (WP_User always truthy)
2. class-flame-builder.php:1529 — nullCoalesce.offset (`$node['value']` guaranteed)
3. class-job-worker.php:379 — identical.alwaysFalse (`ini_get('memory_limit')` returns string)
4. class-memcached-cache.php:96 — nullCoalesce.offset (`explode` returns non-empty-list)
5. class-partition-reader.php:65 — docblock loosened (see above)
6. class-partition-reader.php:66 — same
7. class-remote-manager.php:73 — docblock loosened (see above)
8. class-remote-source.php:595 — identical.alwaysFalse (`wp_json_encode` non-empty-string|false)
9. class-request-builder.php:356 — nullCoalesce.offset (explode non-empty-list)
10. class-request-builder.php:544 — docblock loosened (see above)
11. class-server-registry.php:450 — function.alreadyNarrowedType (`is_string` on string)
12. class-stream-merger.php:523 — nullCoalesce.offset (position() typed shape)
13. class-stream-merger.php:524 — same
14-17. class-perf-requests-controller.php:209,213,214,215 — nullCoalesce.offset
       (parse_flame_index returns typed shape after is_array narrowing)
18. class-requests-stream-controller.php:54 — nullCoalesce.offset (already !empty checked)
19. class-workers-controller.php:180 — nullCoalesce.offset (enumerate_offsetlog_rows typed shape)

## composer.json: pre-commit + commit-msg hooks

App already had `@commitlint/cli`, `@commitlint/config-conventional`,
`lint-staged` devDeps, plus `commitlint.config.js` and `lint-staged`
in package.json — but nothing wired the hooks into `.git/hooks/`.
Now wired via cghooks `extra.hooks`:

    "pre-commit": "./node_modules/.bin/lint-staged"
    "commit-msg": "cat $1 | ./node_modules/.bin/commitlint"

Matches substrate + pyrobase coverage (pre-commit + commit-msg + pre-push).

## Verification

- phpstan: 0 errors
- phpunit: 1828 tests pass

+++ includes/class-partition-reader.php
@@ -52,8 +52,9 @@ class Partition_Reader {
 	/**
 	 * Set the next read position.
 	 *
-	 * @param string|array{segment_id:int,offset:int} $position 'start' / 'recent' /
-	 *                                                          'end' / explicit position.
+	 * @param string|array{segment_id?:int,offset?:int} $position 'start' / 'recent' /
+	 *                                                            'end' / explicit position.
+	 *                                                            Missing array keys default to 0.
 	 */
 	public function next_offset( string|array $position ): void {
 		$this->close();
@@ -62,8 +63,8 @@ class Partition_Reader {
 		$this->at_eof          = false;
 
 		if ( \is_array( $position ) ) {
-			$this->segment_id = (int) ( $position['segment_id'] ?? 0 );
-			$this->offset     = (int) ( $position['offset'] ?? 0 );
+			$this->segment_id = $position['segment_id'] ?? 0;
+			$this->offset     = $position['offset'] ?? 0;
 			if ( $this->offset < -1 ) {
 				$this->offset = -1;
 			}
